+++ changelog.md
@@ -8,6 +8,7 @@
 - 14:15 - Frontend: "User profile" page translated.
 - 16:00 - Frontend: "Folders" page translated.
 - 16:35 - Frontend: "Bookmarks" page translated.
+- 17:15 - Backend: added Redis.
 
 ## 20.04.2023, Thu